Split canvas focus and support views
