+++ supacode/Features/Canvas/Views/CanvasView.swift
@@ -3,8 +3,8 @@ import Sharing
 import SwiftUI
 
 struct CanvasView: View {
-  @Environment(CommandKeyObserver.self) private var commandKeyObserver
-  @Environment(\.resolvedKeybindings) private var resolvedKeybindings
+  @Environment(CommandKeyObserver.self) var commandKeyObserver
+  @Environment(\.resolvedKeybindings) var resolvedKeybindings
 
   let terminalManager: WorktreeTerminalManager
   /// Per-repo display titles resolved by the parent reducer. Used to
@@ -22,46 +22,46 @@ struct CanvasView: View {
   /// Reports whether a card is currently expanded in place, so the parent can
   /// give the window toolbar a matching scrim (it can't be covered from here).
   var onExpandedChange: (Bool) -> Void = { _ in }
-  @State private var layoutStore = CanvasLayoutStore()
-  @Shared(.repositoryAppearances) private var repositoryAppearances
+  @State var layoutStore = CanvasLayoutStore()
+  @Shared(.repositoryAppearances) var repositoryAppearances
 
-  @State private var canvasOffset: CGSize = .zero
-  @State private var lastCanvasOffset: CGSize = .zero
-  @State private var canvasScale: CGFloat = 1.0
-  @State private var lastCanvasScale: CGFloat = 1.0
-  @State private var selectionState = CanvasSelectionState()
-  @State private var lastTitleBarTapDate: Date = .distantPast
-  @State private var activeResize: [TerminalTabID: ActiveResize] = [:]
-  @State private var hasPerformedInitialFit = false
-  @State private var hasSeenCanvasCards = false
-  @State private var viewportSize: CGSize = .zero
-  @State private var showsCanvasHelp = false
-  @State private var configReloadCounter = 0
-  @State private var focusViewportAnimationID = 0
+  @State var canvasOffset: CGSize = .zero
+  @State var lastCanvasOffset: CGSize = .zero
+  @State var canvasScale: CGFloat = 1.0
+  @State var lastCanvasScale: CGFloat = 1.0
+  @State var selectionState = CanvasSelectionState()
+  @State var lastTitleBarTapDate: Date = .distantPast
+  @State var activeResize: [TerminalTabID: ActiveResize] = [:]
+  @State var hasPerformedInitialFit = false
+  @State var hasSeenCanvasCards = false
+  @State var viewportSize: CGSize = .zero
+  @State var showsCanvasHelp = false
+  @State var configReloadCounter = 0
+  @State var focusViewportAnimationID = 0
   /// The tab currently expanded in place (near-fullscreen overlay) on canvas,
   /// or nil when no card is expanded.
-  @State private var expandedTabID: TerminalTabID?
+  @State var expandedTabID: TerminalTabID?
 
-  private let minCardWidth: CGFloat = 300
-  private let minCardHeight: CGFloat = 200
-  private let maxCardWidth: CGFloat = 2400
-  private let maxCardHeight: CGFloat = 1600
-  private let titleBarHeight: CGFloat = 28
-  private let cardSpacing: CGFloat = 20
+  let minCardWidth: CGFloat = 300
+  let minCardHeight: CGFloat = 200
+  let maxCardWidth: CGFloat = 2400
+  let maxCardHeight: CGFloat = 1600
+  let titleBarHeight: CGFloat = 28
+  let cardSpacing: CGFloat = 20
   /// Reserved height at the bottom of the viewport for the help button and
   /// layout toolbar so cards don't sit underneath them after auto-fit.
   /// Cards end up shifted upward by half of this amount.
-  private let bottomToolbarReserve: CGFloat = 50
+  let bottomToolbarReserve: CGFloat = 50
   /// Margin kept on every side of a card temporarily expanded to near-fullscreen.
-  private let expandPadding: CGFloat = 40
+  let expandPadding: CGFloat = 40
   /// Shared animation for expand / restore / relayout. Matches the easeInOut
   /// 0.2s that `CanvasCardView` uses to animate `cardSize`, so the canvas
   /// scale/offset stays in lock-step with the card's terminal size refit.
-  private let expandAnimation: Animation = .easeInOut(duration: 0.2)
+  let expandAnimation: Animation = .easeInOut(duration: 0.2)
 
   /// Width of the screen hosting the canvas window, used to scale the default
   /// card size. Falls back to the large-screen reference when unknown.
-  private var hostScreenWidth: CGFloat {
+  var hostScreenWidth: CGFloat {
     (NSApp.keyWindow?.screen ?? NSScreen.main)?.frame.width
       ?? CanvasCardLayout.maxDefaultScreenWidth
   }
@@ -69,7 +69,7 @@ struct CanvasView: View {
   /// Default size for newly created and uniformly arranged cards, scaled to the
   /// host screen so small screens (14") don't zoom out into tiny text while
   /// large screens still get the roomier card.
-  private var adaptiveDefaultCardSize: CGSize {
+  var adaptiveDefaultCardSize: CGSize {
     CanvasCardLayout.adaptiveDefaultSize(forScreenWidth: hostScreenWidth)
   }
 
@@ -232,7 +232,7 @@ struct CanvasView: View {
     .onDisappear { deactivateCanvas() }
   }
 
-  private func showsSelectionShield(for tabID: TerminalTabID) -> Bool {
+  func showsSelectionShield(for tabID: TerminalTabID) -> Bool {
     if commandKeyObserver.isPressed { return true }
     if selectionState.isSelecting { return true }
     if selectionState.isBroadcasting, selectionState.primaryTabID != tabID { return true }
@@ -245,7 +245,7 @@ struct CanvasView: View {
   /// Uses .offset() (not .position()) to avoid parent size proposals
   /// reaching the NSView, keeping terminal grid stable during zoom.
   @ViewBuilder
-  private func cardsLayer(activeStates: [WorktreeTerminalState]) -> some View {
+  func cardsLayer(activeStates: [WorktreeTerminalState]) -> some View {
     // Pin to .topLeading and fill the viewport so each card's `.offset()` keeps
     // the same (0,0) origin it had under GeometryReader — otherwise the scrim's
     // full-size frame would resize the stack and shift the cards' base position.
@@ -282,7 +282,7 @@ struct CanvasView: View {
   }
 
   @ViewBuilder
-  private func cardView(
+  func cardView(
     for tab: TerminalTabItem,
     in state: WorktreeTerminalState,
     activeStates: [WorktreeTerminalState]
@@ -391,7 +391,7 @@ struct CanvasView: View {
 
   // MARK: - Canvas Gestures
 
-  private var canvasPanGesture: some Gesture {
+  var canvasPanGesture: some Gesture {
     DragGesture()
       .onChanged { value in
         canvasOffset = CGSize(
@@ -404,7 +404,7 @@ struct CanvasView: View {
       }
   }
 
-  private var canvasZoomGesture: some Gesture {
+  var canvasZoomGesture: some Gesture {
     MagnifyGesture()
       .onChanged { value in
         let newScale = max(0.25, min(2.0, lastCanvasScale * value.magnification))
@@ -434,7 +434,7 @@ struct CanvasView: View {
   /// Batch-position all cards that don't have stored layouts yet.
   /// Uses a single, consistent column count to avoid overlap between
   /// cards positioned in different passes.
-  private func ensureLayouts(for cardKeys: [String]) {
+  func ensureLayouts(for cardKeys: [String]) {
     let unpositioned = cardKeys.filter { layoutStore.cardLayouts[$0] == nil }
     guard !unpositioned.isEmpty else { return }
 
@@ -461,11 +461,11 @@ struct CanvasView: View {
 
   /// Balanced grid: columns ≈ sqrt(N). No viewport constraint — the canvas
   /// is infinite and fitToView handles zoom.
-  private func gridColumns(for count: Int) -> Int {
+  func gridColumns(for count: Int) -> Int {
     max(1, Int(ceil(sqrt(Double(count)))))
   }
 
-  private func gridPosition(index: Int, columns: Int, cardSize: CGSize) -> CGPoint {
+  func gridPosition(index: Int, columns: Int, cardSize: CGSize) -> CGPoint {
     let cardW = cardSize.width
     let cardH = cardSize.height + titleBarHeight
     let row = index / columns
@@ -478,7 +478,7 @@ struct CanvasView: View {
 
   /// Compute effective center and size accounting for resize only (not drag).
   /// Drag is applied separately via `.offset()` to avoid layout passes.
-  private func resizedFrame(
+  func resizedFrame(
     for tabID: TerminalTabID,
     baseLayout: CanvasCardLayout
   ) -> (center: CGPoint, size: CGSize) {
@@ -504,24 +504,24 @@ struct CanvasView: View {
     return (CGPoint(x: centerX, y: centerY), CGSize(width: width, height: height))
   }
 
-  private func screenPosition(for canvasCenter: CGPoint) -> CGPoint {
+  func screenPosition(for canvasCenter: CGPoint) -> CGPoint {
     CGPoint(
       x: canvasCenter.x * canvasScale + canvasOffset.width,
       y: canvasCenter.y * canvasScale + canvasOffset.height
     )
   }
 
-  private func clampWidth(_ width: CGFloat) -> CGFloat {
+  func clampWidth(_ width: CGFloat) -> CGFloat {
     max(minCardWidth, min(maxCardWidth, width))
   }
 
-  private func clampHeight(_ height: CGFloat) -> CGFloat {
+  func clampHeight(_ height: CGFloat) -> CGFloat {
     max(minCardHeight, min(maxCardHeight, height))
   }
 
   // MARK: - Organize & Fit
 
-  private func collectCardKeys(from states: [WorktreeTerminalState]) -> [String] {
+  func collectCardKeys(from states: [WorktreeTerminalState]) -> [String] {
     states.flatMap { state in
       state.tabManager.tabs.compactMap { tab in
         state.surfaceView(for: tab.id) != nil ? tab.id.rawValue.uuidString : nil
@@ -529,7 +529,7 @@ struct CanvasView: View {
     }
   }
 
-  private func collectVisibleTabIDs(from states: [WorktreeTerminalState]) -> [TerminalTabID] {
+  func collectVisibleTabIDs(from states: [WorktreeTerminalState]) -> [TerminalTabID] {
     states.flatMap { state in
       state.tabManager.tabs.compactMap { tab in
         state.surfaceView(for: tab.id) != nil ? tab.id : nil
@@ -537,7 +537,7 @@ struct CanvasView: View {
     }
   }
 
-  private func collectFocusCandidates(from states: [WorktreeTerminalState]) -> [CanvasFocusCandidate] {
+  func collectFocusCandidates(from states: [WorktreeTerminalState]) -> [CanvasFocusCandidate] {
     states.flatMap { state in
       state.tabManager.tabs.compactMap { tab in
         state.surfaceView(for: tab.id) != nil
@@ -548,7 +548,7 @@ struct CanvasView: View {
   }
 
   /// Reset all card positions to a clean grid layout (uniform sizes).
-  private func organizeCards() {
+  func organizeCards() {
     let keys = collectCardKeys(from: terminalManager.activeWorktreeStates)
     let columns = gridColumns(for: keys.count)
     let cardSize = adaptiveDefaultCardSize
@@ -565,7 +565,7 @@ struct CanvasView: View {
   /// Arrange cards using MaxRects-BSSF bin packing. Preserves each card's
   /// current size and finds a compact layout whose aspect ratio matches
   /// the viewport.
-  private func arrangeCards() {
+  func arrangeCards() {
     let keys = collectCardKeys(from: terminalManager.activeWorktreeStates)
     guard !keys.isEmpty, viewportSize.width > 0, viewportSize.height > 0 else { return }
 
@@ -584,7 +584,7 @@ struct CanvasView: View {
 
   /// Arrange cards (preserving sizes) and refit the viewport, animated.
   /// Shared by the toolbar button and the keyboard shortcut.
-  private func arrangeCardsWithFit() {
+  func arrangeCardsWithFit() {
     withAnimation(.easeInOut(duration: 0.2)) {
       cancelExpandForRelayout()
       arrangeCards()
@@ -594,7 +594,7 @@ struct CanvasView: View {
 
   /// Organize cards into a uniform grid and refit the viewport, animated.
   /// Shared by the toolbar button and the keyboard shortcut.
-  private func organizeCardsWithFit() {
+  func organizeCardsWithFit() {
     withAnimation(.easeInOut(duration: 0.2)) {
       cancelExpandForRelayout()
       organizeCards()
@@ -603,7 +603,7 @@ struct CanvasView: View {
   }
 
   /// Adjust scale and offset so all cards fit within the viewport.
-  private func fitToView(canvasSize: CGSize) {
+  func fitToView(canvasSize: CGSize) {
     guard canvasSize.width > 0, canvasSize.height > 0 else { return }
 
     let keys = collectCardKeys(from: terminalManager.activeWorktreeStates)
@@ -645,13 +645,13 @@ struct CanvasView: View {
   }
 
   /// Remove stored layouts for tabs that no longer exist.
-  private func cleanStaleLayouts() {
+  func cleanStaleLayouts() {
     let visibleKeys = Set(collectCardKeys(from: terminalManager.activeWorktreeStates))
     guard !visibleKeys.isEmpty || hasSeenCanvasCards else { return }
     layoutStore.prune(to: visibleKeys)
   }
 
-  private var canvasHelpButton: some View {
+  var canvasHelpButton: some View {
     Button {
       showsCanvasHelp.toggle()
     } label: {
@@ -667,7 +667,7 @@ struct CanvasView: View {
     .padding()
   }
 
-  private var canvasHelpContent: some View {
+  var canvasHelpContent: some View {
     let expandShortcut = AppShortcuts.display(
       for: AppShortcuts.CommandID.expandCanvasCard,
       in: resolvedKeybindings
@@ -699,7 +699,7 @@ struct CanvasView: View {
     .frame(width: 320, alignment: .leading)
   }
 
-  private func canvasHelpRow(icon: String, title: String, detail: String) -> some View {
+  func canvasHelpRow(icon: String, title: String, detail: String) -> some View {
     HStack(alignment: .firstTextBaseline, spacing: 10) {
       Image(systemName: icon)
         .foregroundStyle(.secondary)
@@ -715,7 +715,7 @@ struct CanvasView: View {
     }
   }
 
-  private var canvasToolbar: some View {
+  var canvasToolbar: some View {
     HStack(spacing: 8) {
       if selectionState.isBroadcasting {
         Label(
@@ -776,7 +776,7 @@ struct CanvasView: View {
     .padding()
   }
 
-  private func zIndex(for tabID: TerminalTabID, cardKey: String) -> Double {
+  func zIndex(for tabID: TerminalTabID, cardKey: String) -> Double {
     let base = layoutStore.zIndex(for: cardKey)
     if selectionState.primaryTabID == tabID {
       return 10_000 + base
@@ -789,7 +789,7 @@ struct CanvasView: View {
 
   // MARK: - Drag
 
-  private func commitDrag(for cardKey: String, translation: CGSize) {
+  func commitDrag(for cardKey: String, translation: CGSize) {
     if var layout = layoutStore.cardLayouts[cardKey] {
       layout.position.x += translation.width
       layout.position.y += translation.height
@@ -799,7 +799,7 @@ struct CanvasView: View {
 
   // MARK: - Resize
 
-  private func commitResize(for tabID: TerminalTabID, cardKey: String, surfaces: [GhosttySurfaceView]) {
+  func commitResize(for tabID: TerminalTabID, cardKey: String, surfaces: [GhosttySurfaceView]) {
     guard activeResize[tabID] != nil else { return }
     if var layout = layoutStore.cardLayouts[cardKey] {
       let resized = resizedFrame(for: tabID, baseLayout: layout)
@@ -818,7 +818,7 @@ struct CanvasView: View {
     }
   }
 
-  private func selectAllCards() {
+  func selectAllCards() {
     let activeStates = terminalManager.activeWorktreeStates
     let allTabIDs = collectVisibleTabIDs(from: activeStates)
     guard allTabIDs.count > 1 else { return }
@@ -829,7 +829,7 @@ struct CanvasView: View {
 
   // MARK: - Selection and Focus
 
-  private func focusSingleCard(
+  func focusSingleCard(
     _ tabID: TerminalTabID,
     states: [WorktreeTerminalState]
   ) {
@@ -841,7 +841,7 @@ struct CanvasView: View {
 
   // MARK: - Expand In Place
 
-  private var expandMetrics: CanvasExpandGeometry.Metrics {
+  var expandMetrics: CanvasExpandGeometry.Metrics {
     CanvasExpandGeometry.Metrics(
       padding: expandPadding,
       bottomReserve: bottomToolbarReserve,
@@ -852,14 +852,14 @@ struct CanvasView: View {
 
   /// Screen-space center for a fully expanded card: horizontally centered and
   /// within the toolbar-adjusted viewport. Independent of canvas pan/zoom.
-  private var expandedScreenCenter: CGPoint {
+  var expandedScreenCenter: CGPoint {
     CGPoint(x: viewportSize.width / 2, y: (viewportSize.height - bottomToolbarReserve) / 2)
   }
 
   /// A card's normal (non-expanded) on-screen frame, following the canvas
   /// pan/zoom and any in-progress resize. This is the `progress = 0` endpoint of
   /// the expand magic-move.
-  private func nonExpandedGeometry(
+  func nonExpandedGeometry(
     for tabID: TerminalTabID,
     baseLayout: CanvasCardLayout
   ) -> CardScreenGeometry {
@@ -874,7 +874,7 @@ struct CanvasView: View {
   /// The full-viewport expanded frame at scale 1 — the `progress = 1` endpoint.
   /// Independent of the canvas transform, so it covers the viewport regardless
   /// of the (frozen) background.
-  private func expandedGeometry() -> CardScreenGeometry {
+  func expandedGeometry() -> CardScreenGeometry {
     CardScreenGeometry(
       size: CanvasExpandGeometry.expandedSize(viewport: viewportSize, metrics: expandMetrics),
       center: expandedScreenCenter,
@@ -884,7 +884,7 @@ struct CanvasView: View {
 
   /// Toggle expand/restore for a card — used by the title-bar button and the
   /// title-bar double-click.
-  private func toggleExpand(_ tabID: TerminalTabID, states: [WorktreeTerminalState]) {
+  func toggleExpand(_ tabID: TerminalTabID, states: [WorktreeTerminalState]) {
     if expandedTabID == tabID {
       collapseExpand()
     } else {
@@ -894,715 +894,11 @@ struct CanvasView: View {
 
   /// Toggle expand/restore for the focused (primary) card. Used by the keyboard
   /// shortcut and the command palette, which target whichever card is focused.
-  private func toggleExpandFocusedCard() {
+  func toggleExpandFocusedCard() {
     if expandedTabID != nil {
       collapseExpand()
     } else if let tabID = selectionState.primaryTabID {
       expandCard(tabID, states: terminalManager.activeWorktreeStates)
     }
-  }
-
-  /// Run a reducer-driven canvas command (from the command palette) and clear
-  /// the one-shot request.
-  private func fulfillCommandRequest(_ request: CanvasCommandRequest?) {
-    guard let request else { return }
-    switch request.command {
-    case .toggleExpand:
-      toggleExpandFocusedCard()
-    case .arrange:
-      arrangeCardsWithFit()
-    case .organize:
-      organizeCardsWithFit()
-    case .selectAll:
-      selectAllCards()
-    }
-    onCommandConsumed(request.id)
-  }
-
-  /// Expand a card in place: raise it to the top, then flip `expandedTabID`
-  /// inside one `withAnimation` so the card's size, center, and scale all
-  /// interpolate together from its in-canvas frame to the full viewport at
-  /// scale 1 — a magic-move from where the card actually sits. The canvas
-  /// transform is left untouched, so every other card stays exactly where it
-  /// was, behind a dimming scrim.
-  private func expandCard(_ tabID: TerminalTabID, states: [WorktreeTerminalState]) {
-    guard viewportSize.width > 0, viewportSize.height > 0,
-      layoutStore.cardLayouts[tabID.rawValue.uuidString] != nil
-    else { return }
-    focusSingleCard(tabID, states: states)
-    withAnimation(expandAnimation) {
-      expandedTabID = tabID
-    }
-  }
-
-  /// Restore the expanded card back into the (unchanged) canvas.
-  private func collapseExpand() {
-    guard expandedTabID != nil else { return }
-    withAnimation(expandAnimation) {
-      expandedTabID = nil
-    }
-  }
-
-  /// Drop expand state without animation — used right before a relayout
-  /// (Arrange/Organize) takes over the canvas.
-  private func cancelExpandForRelayout() {
-    expandedTabID = nil
-  }
-
-  private func fulfillPendingFocusRequest(
-    _ request: CanvasFocusRequest?,
-    states: [WorktreeTerminalState]
-  ) {
-    guard let request else { return }
-    let tabID = CanvasFocusResolver.resolve(
-      request: request,
-      candidates: collectFocusCandidates(from: states),
-      currentPrimaryTabID: selectionState.primaryTabID
-    )
-    guard let tabID else { return }
-    focusSingleCard(tabID, states: states)
-    focusViewport(on: tabID)
-    onFocusRequestConsumed(request.id)
-  }
-
-  private func focusViewport(on tabID: TerminalTabID) {
-    guard viewportSize.width > 0, viewportSize.height > 0 else { return }
-    let cardKey = tabID.rawValue.uuidString
-    guard let layout = layoutStore.cardLayouts[cardKey] else { return }
-
-    let horizontalPadding: CGFloat = 80
-    let verticalPadding: CGFloat = 80 + bottomToolbarReserve
-    let availableWidth = max(1, viewportSize.width - horizontalPadding)
-    let availableHeight = max(1, viewportSize.height - verticalPadding)
-    let targetScale = max(
-      0.25,
-      min(
-        1.25,
-        min(
-          availableWidth / layout.size.width,
-          availableHeight / (layout.size.height + titleBarHeight)
-        )
-      )
-    )
-    let targetOffset = CGSize(
-      width: viewportSize.width / 2 - layout.position.x * targetScale,
-      height: (viewportSize.height - bottomToolbarReserve) / 2 - layout.position.y * targetScale
-    )
-    canvasScale = targetScale
-    canvasOffset = targetOffset
-    lastCanvasScale = targetScale
-    lastCanvasOffset = targetOffset
-    focusViewportAnimationID &+= 1
-  }
-
-  private func handleSelectionShieldTap(
-    _ tabID: TerminalTabID,
-    surfaceState _: WorktreeTerminalState,
-    states: [WorktreeTerminalState]
-  ) {
-    let cmdHeld = NSEvent.modifierFlags.contains(.command)
-    mutateSelection(states: states) { state in
-      if cmdHeld {
-        state.toggleSelection(tabID)
-      } else if state.isBroadcasting, state.selectedTabIDs.contains(tabID) {
-        state.setPrimary(tabID)
-      } else {
-        state.focusSingle(tabID)
-      }
-    }
-  }
-
-  private func clearSelection(states: [WorktreeTerminalState]) {
-    mutateSelection(states: states) { state in
-      state.clear()
-    }
-  }
-
-  private func pruneSelection(
-    previousOrder: [TerminalTabID],
-    currentOrder: [TerminalTabID],
-    states: [WorktreeTerminalState]
-  ) {
-    let previousPrimaryTabID = selectionState.primaryTabID
-    selectionState.pruneAutoAdvancingPrimary(previousOrder: previousOrder, currentOrder: currentOrder)
-    syncPrimaryFocus(from: previousPrimaryTabID, to: selectionState.primaryTabID, states: states)
-    syncBroadcastCallbacks(states: states)
-  }
-
-  private func mutateSelection(
-    states: [WorktreeTerminalState],
-    mutation: (inout CanvasSelectionState) -> Void
-  ) {
-    let previousPrimaryTabID = selectionState.primaryTabID
-    mutation(&selectionState)
-    selectionState.prune(to: Set(collectVisibleTabIDs(from: states)))
-    syncPrimaryFocus(from: previousPrimaryTabID, to: selectionState.primaryTabID, states: states)
-    syncBroadcastCallbacks(states: states)
-  }
-
-  private func syncPrimaryFocus(
-    from previousTabID: TerminalTabID?,
-    to newTabID: TerminalTabID?,
-    states: [WorktreeTerminalState]
-  ) {
-    if let previousTabID, previousTabID != newTabID {
-      unfocusTab(previousTabID, states: states)
-    }
-
-    guard let newTabID,
-      let ownerState = states.first(where: { $0.surfaceView(for: newTabID) != nil }),
-      let surfaceView = ownerState.surfaceView(for: newTabID)
-    else {
-      terminalManager.canvasFocusedWorktreeID = nil
-      onFocusedWorktreeChanged(nil)
-      return
-    }
-
-    layoutStore.moveToFront(newTabID.rawValue.uuidString)
-    ownerState.tabManager.selectTab(newTabID)
-    terminalManager.canvasFocusedWorktreeID = ownerState.worktreeID
-    onFocusedWorktreeChanged(ownerState.worktreeID)
-    surfaceView.focusDidChange(true)
-    surfaceView.requestFocus()
-  }
-
-  private func unfocusTab(_ tabID: TerminalTabID, states: [WorktreeTerminalState]) {
-    guard let state = states.first(where: { $0.surfaceView(for: tabID) != nil }) else { return }
-    for surface in state.splitTree(for: tabID).leaves() {
-      surface.focusDidChange(false)
-    }
-  }
-
-  private func syncBroadcastCallbacks(states: [WorktreeTerminalState]) {
-    clearBroadcastCallbacks(states: states)
-
-    guard selectionState.isBroadcasting,
-      let primaryTabID = selectionState.primaryTabID,
-      let primaryState = terminalManager.stateContaining(tabId: primaryTabID)
-    else {
-      return
-    }
-
-    let selectedTabIDs = selectionState.selectedTabIDs
-    let beginBroadcast = { selectionState.beginBroadcastInteractionIfNeeded() }
-    for primarySurface in primaryState.splitTree(for: primaryTabID).leaves() {
-      primarySurface.onCommittedText = { [terminalManager, selectedTabIDs, primaryTabID, beginBroadcast] text in
-        Task { @MainActor in
-          beginBroadcast()
-          terminalManager.broadcastCommittedText(text, from: primaryTabID, to: selectedTabIDs)
-        }
-      }
-      primarySurface.onMirroredKey = {
-        [terminalManager, selectedTabIDs, primaryTabID, beginBroadcast] mirroredKey in
-        Task { @MainActor in
-          beginBroadcast()
-          terminalManager.broadcastMirroredKey(mirroredKey, from: primaryTabID, to: selectedTabIDs)
-        }
-      }
-    }
-  }
-
-  private func clearBroadcastCallbacks(states: [WorktreeTerminalState]) {
-    for state in states {
-      for tab in state.tabManager.tabs {
-        for surface in state.splitTree(for: tab.id).leaves() {
-          surface.onCommittedText = nil
-          surface.onMirroredKey = nil
-        }
-      }
-    }
-  }
-
-  // MARK: - Occlusion
-
-  private func activateCanvas() {
-    cleanStaleLayouts()
-
-    let activeStates = terminalManager.activeWorktreeStates
-
-    // Mark all states as canvas-managed so that tree updates (e.g. split
-    // creation) don't trigger applySurfaceActivity with stale normal-mode
-    // window visibility, which would occlude every surface.
-    for state in activeStates {
-      state.isCanvasManaged = true
-    }
-
-    // Auto-focus the card that was active before entering canvas.
-    if let selectedID = terminalManager.selectedWorktreeID,
-      let state = activeStates.first(where: { $0.worktreeID == selectedID }),
-      let tabID = state.tabManager.selectedTabId
-    {
-      selectionState.focusSingle(tabID)
-      syncPrimaryFocus(from: nil, to: tabID, states: activeStates)
-    } else {
-      selectionState.clear()
-      syncBroadcastCallbacks(states: activeStates)
-    }
-
-    for state in activeStates {
-      state.setAllSurfacesOccluded()
-    }
-    // Un-occlude all surfaces visible on canvas (including split panes)
-    for state in activeStates {
-      for tab in state.tabManager.tabs {
-        for surface in state.splitTree(for: tab.id).leaves() {
-          surface.setOcclusion(true)
-        }
-      }
-    }
-  }
-
-  private func deactivateCanvas() {
-    expandedTabID = nil
-    let activeStates = terminalManager.activeWorktreeStates
-    for state in activeStates {
-      state.isCanvasManaged = false
-    }
-    clearBroadcastCallbacks(states: activeStates)
-    selectionState.clear()
-    // Don't occlude surfaces here. In SwiftUI's if/else view swap,
-    // onAppear fires before onDisappear, so occluding here would undo
-    // WorktreeTerminalTabsView.onAppear's syncFocus() and cause blank
-    // surfaces. Cleanup of non-selected worktrees is handled by
-    // setSelectedWorktreeID in the async exit flow.
-  }
-
-  /// Looks up the user-pinned `RepositoryAppearance` for a given repo
-  /// root URL by deriving the canonical `Repository.ID` (the
-  /// path-policy-normalized path string) and querying the @Shared
-  /// dict. Returns `.empty` when no entry exists, which keeps cards
-  /// visually identical to before the appearance feature shipped.
-  private func appearance(for repositoryRootURL: URL) -> RepositoryAppearance {
-    let id = repositoryID(for: repositoryRootURL)
-    return repositoryAppearances[id] ?? .empty
-  }
-
-  /// Resolves the user-defined display title for the repo at this root
-  /// URL, falling back to `Repository.name(for:)` (folder name) when no
-  /// custom title was set. Reads from the static dictionary populated
-  /// by the parent reducer — no per-call `@Shared` subscription on the
-  /// canvas hot path.
-  private func repositoryDisplayName(for repositoryRootURL: URL) -> String {
-    let id = repositoryID(for: repositoryRootURL)
-    return repositoryCustomTitles[id] ?? Repository.name(for: repositoryRootURL)
-  }
-
-  /// Mirrors the same path normalization the `Repository.ID` is built
-  /// from, so dict lookups match what the reducer stores.
-  private func repositoryID(for repositoryRootURL: URL) -> Repository.ID {
-    PathPolicy.normalizePath(
-      repositoryRootURL.path(percentEncoded: false), resolvingSymlinks: true
-    ) ?? repositoryRootURL.path(percentEncoded: false)
-  }
-}
-
-private struct ActiveResize {
-  let edge: CanvasCardView.CardResizeEdge
-  var translation: CGSize
-}
-
-// MARK: - Scroll Container
-
-/// Wraps SwiftUI content in an NSView whose `scrollWheel` override catches
-/// unhandled scroll-wheel events and translates them into canvas-offset changes.
-/// Focused terminals consume their own scroll events (they don't call super),
-/// so only events over empty space or unfocused cards reach this container.
-private struct CanvasScrollContainer<Content: View>: NSViewRepresentable {
-  @Binding var offset: CGSize
-  @Binding var lastOffset: CGSize
-  @Binding var scale: CGFloat
-  @Binding var lastScale: CGFloat
-  var isInteractionEnabled: Bool
-  @ViewBuilder var content: Content
-
-  func makeCoordinator() -> CanvasScrollCoordinator {
-    CanvasScrollCoordinator()
-  }
-
-  func makeNSView(context: Context) -> CanvasScrollContainerView {
-    let container = CanvasScrollContainerView()
-    let hosting = NSHostingView(rootView: content)
-    hosting.translatesAutoresizingMaskIntoConstraints = false
-    container.addSubview(hosting)
-    NSLayoutConstraint.activate([
-      hosting.topAnchor.constraint(equalTo: container.topAnchor),
-      hosting.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-      hosting.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-      hosting.trailingAnchor.constraint(equalTo: container.trailingAnchor),
-    ])
-    container.scrollCoordinator = context.coordinator
-    return container
-  }
-
-  func updateNSView(_ nsView: CanvasScrollContainerView, context: Context) {
-    context.coordinator.offset = $offset
-    context.coordinator.lastOffset = $lastOffset
-    context.coordinator.scale = $scale
-    context.coordinator.lastScale = $lastScale
-    nsView.isInteractionEnabled = isInteractionEnabled
-    if let hosting = nsView.subviews.first as? NSHostingView<Content> {
-      hosting.rootView = content
-    }
-  }
-}
-
-private class CanvasScrollCoordinator {
-  var offset: Binding<CGSize> = .constant(.zero)
-  var lastOffset: Binding<CGSize> = .constant(.zero)
-  var scale: Binding<CGFloat> = .constant(1.0)
-  var lastScale: Binding<CGFloat> = .constant(1.0)
-
-  func handleScroll(deltaX: CGFloat, deltaY: CGFloat) {
-    let current = offset.wrappedValue
-    let newOffset = CGSize(
-      width: current.width + deltaX,
-      height: current.height + deltaY
-    )
-    offset.wrappedValue = newOffset
-    lastOffset.wrappedValue = newOffset
-  }
-
-  func handleZoom(deltaY: CGFloat, anchor: CGPoint, isPrecise: Bool) {
-    let result = CanvasZoomMath.zoom(
-      currentScale: scale.wrappedValue,
-      currentOffset: offset.wrappedValue,
-      deltaY: deltaY,
-      anchor: anchor,
-      isPrecise: isPrecise
-    )
-    scale.wrappedValue = result.scale
-    lastScale.wrappedValue = result.scale
-    offset.wrappedValue = result.offset
-    lastOffset.wrappedValue = result.offset
-  }
-
-  func setOffset(_ newOffset: CGSize) {
-    offset.wrappedValue = newOffset
-    lastOffset.wrappedValue = newOffset
-  }
-}
-
-/// Pure zoom math, extracted for testability.
-enum CanvasZoomMath {
-  static let minScale: CGFloat = 0.25
-  static let maxScale: CGFloat = 2.0
-
-  struct Result: Equatable {
-    let scale: CGFloat
-    let offset: CGSize
-  }
-
-  /// Compute the new scale and offset for a Cmd+wheel zoom step.
-  /// Keeps the canvas point under `anchor` fixed under the cursor:
-  /// `screen = canvas * scale + offset` ⇒ `canvas = (anchor - offset) / scale`.
-  static func zoom(
-    currentScale: CGFloat,
-    currentOffset: CGSize,
-    deltaY: CGFloat,
-    anchor: CGPoint,
-    isPrecise: Bool
-  ) -> Result {
-    let sensitivity: CGFloat = isPrecise ? 0.0025 : 0.005
-    let factor = exp(deltaY * sensitivity)
-    let newScale = max(minScale, min(maxScale, currentScale * factor))
-    guard newScale != currentScale else {
-      return Result(scale: currentScale, offset: currentOffset)
-    }
-    let canvasX = (anchor.x - currentOffset.width) / currentScale
-    let canvasY = (anchor.y - currentOffset.height) / currentScale
-    let newOffset = CGSize(
-      width: anchor.x - canvasX * newScale,
-      height: anchor.y - canvasY * newScale
-    )
-    return Result(scale: newScale, offset: newOffset)
-  }
-}
-
-private class CanvasScrollContainerView: NSView {
-  var scrollCoordinator: CanvasScrollCoordinator?
-  /// When false (a card is expanded), the container ignores scroll/zoom/
-  /// middle-drag so the canvas can't pan or zoom behind the expanded card.
-  var isInteractionEnabled = true {
-    didSet {
-      guard !isInteractionEnabled, oldValue else { return }
-      if isMiddlePanning { endMiddlePan() }
-      isPanning = false
-    }
-  }
-
-  /// Whether the container is actively redirecting scroll events to canvas
-  /// panning (as opposed to the brief bounce period after a gesture ends).
-  private var isPanning = false
-  private var scrollMonitor: Any?
-  /// Brief delay after finger-up to wait for momentum events.
-  private var momentumTimer: Timer?
-  /// Grace period after a pan gesture ends. A follow-up gesture that begins
-  /// during this window is still treated as canvas panning, even if the
-  /// cursor now sits on a focused terminal.
-  private var bounceTimer: Timer?
-
-  // MARK: - Middle-click pan
-  private var middleButtonMonitor: Any?
-  private var isMiddlePanning = false
-  private var middlePanStartLocation: NSPoint = .zero
-  private var middlePanStartOffset: CGSize = .zero
-  private var hasPushedPanCursor = false
-
-  override func scrollWheel(with event: NSEvent) {
-    guard isInteractionEnabled else {
-      super.scrollWheel(with: event)
-      return
-    }
-    if handleZoomEventIfNeeded(event) { return }
-    if event.phase == .began {
-      startPanning()
-    }
-    if event.phase == .began || event.phase == .changed || event.phase == .mayBegin || event.momentumPhase != [] {
-      scrollCoordinator?.handleScroll(deltaX: event.scrollingDeltaX, deltaY: event.scrollingDeltaY)
-      return
-    }
-    super.scrollWheel(with: event)
-  }
-
-  /// If the event is a Cmd+scroll, route it to canvas zoom and report `true`.
-  /// Used by both the direct `scrollWheel` override and the local monitor so
-  /// pressing Cmd mid-gesture switches behavior immediately.
-  fileprivate func handleZoomEventIfNeeded(_ event: NSEvent) -> Bool {
-    guard isInteractionEnabled else { return false }
-    guard event.modifierFlags.contains(.command), event.scrollingDeltaY != 0 else { return false }
-    let viewLocation = convert(event.locationInWindow, from: nil)
-    let anchor = CGPoint(x: viewLocation.x, y: bounds.height - viewLocation.y)
-    scrollCoordinator?.handleZoom(
-      deltaY: event.scrollingDeltaY,
-      anchor: anchor,
-      isPrecise: event.hasPreciseScrollingDeltas
-    )
-    return true
-  }
-
-  // MARK: - Pan lifecycle
-
-  private func startPanning() {
-    isPanning = true
-    momentumTimer?.invalidate()
-    momentumTimer = nil
-    bounceTimer?.invalidate()
-    bounceTimer = nil
-    guard scrollMonitor == nil else { return }
-    installMonitor()
-  }
-
-  private func installMonitor() {
-    scrollMonitor = NSEvent.addLocalMonitorForEvents(matching: .scrollWheel) { [weak self] event in
-      guard let self, event.window === self.window else { return event }
-      guard self.isInteractionEnabled else { return event }
-
-      // Cmd toggled mid-gesture — switch to zoom for this event.
-      if self.handleZoomEventIfNeeded(event) { return nil }
-
-      // --- New gesture ------------------------------------------------
-      if event.phase == .began {
-        if self.isPanning {
-          // Already panning (edge case). Let normal dispatch decide.
-          return event
-        }
-        // Within the bounce window — treat as a continuation of panning.
-        self.startPanning()
-        self.scrollCoordinator?.handleScroll(
-          deltaX: event.scrollingDeltaX,
-          deltaY: event.scrollingDeltaY
-        )
-        return nil
-      }
-
-      // Only intercept while actively panning (not during bounce).
-      guard self.isPanning else { return event }
-
-      // --- Ongoing gesture / momentum --------------------------------
-      self.momentumTimer?.invalidate()
-      self.momentumTimer = nil
-
-      if event.phase == .changed || event.momentumPhase != [] {
-        self.scrollCoordinator?.handleScroll(
-          deltaX: event.scrollingDeltaX,
-          deltaY: event.scrollingDeltaY
-        )
-      }
-
-      // Finger lifted — momentum may follow shortly.
-      if event.phase == .ended || event.phase == .cancelled {
-        self.momentumTimer = Timer.scheduledTimer(
-          withTimeInterval: 0.1, repeats: false
-        ) { [weak self] _ in
-          MainActor.assumeIsolated { self?.enterBounce() }
-        }
-      }
-
-      // Momentum finished.
-      if event.momentumPhase == .ended || event.momentumPhase == .cancelled {
-        self.enterBounce()
-      }
-
-      return nil
-    }
-  }
-
-  /// Transition from active panning to the bounce (grace) period.
-  /// The monitor stays alive so a quick follow-up gesture resumes panning.
-  private func enterBounce() {
-    isPanning = false
-    momentumTimer?.invalidate()
-    momentumTimer = nil
-    bounceTimer?.invalidate()
-    bounceTimer = Timer.scheduledTimer(
-      withTimeInterval: 0.3, repeats: false
-    ) { [weak self] _ in
-      MainActor.assumeIsolated { self?.tearDownMonitor() }
-    }
-  }
-
-  private func tearDownMonitor() {
-    isPanning = false
-    momentumTimer?.invalidate()
-    momentumTimer = nil
-    bounceTimer?.invalidate()
-    bounceTimer = nil
-    if let monitor = scrollMonitor {
-      scrollMonitor = nil
-      DispatchQueue.main.async { MainActor.assumeIsolated { NSEvent.removeMonitor(monitor) } }
-    }
-  }
-
-  // MARK: - Middle-click pan
-
-  override func viewDidMoveToWindow() {
-    super.viewDidMoveToWindow()
-    if window != nil {
-      installMiddleButtonMonitor()
-    } else {
-      tearDownMiddleButtonMonitor()
-    }
-  }
-
-  private func installMiddleButtonMonitor() {
-    guard middleButtonMonitor == nil else { return }
-    let mask: NSEvent.EventTypeMask = [.otherMouseDown, .otherMouseDragged, .otherMouseUp]
-    middleButtonMonitor = NSEvent.addLocalMonitorForEvents(matching: mask) { [weak self] event in
-      guard let self, event.window === self.window, event.buttonNumber == 2 else { return event }
-      guard self.isInteractionEnabled else { return event }
-
-      switch event.type {
-      case .otherMouseDown:
-        let location = self.convert(event.locationInWindow, from: nil)
-        guard self.bounds.contains(location) else { return event }
-        self.beginMiddlePan(at: event.locationInWindow)
-        return nil
-      case .otherMouseDragged:
-        guard self.isMiddlePanning else { return event }
-        self.updateMiddlePan(to: event.locationInWindow)
-        return nil
-      case .otherMouseUp:
-        guard self.isMiddlePanning else { return event }
-        self.endMiddlePan()
-        return nil
-      default:
-        return event
-      }
-    }
-  }
-
-  private func beginMiddlePan(at windowLocation: NSPoint) {
-    isMiddlePanning = true
-    middlePanStartLocation = windowLocation
-    middlePanStartOffset = scrollCoordinator?.offset.wrappedValue ?? .zero
-    if !hasPushedPanCursor {
-      NSCursor.closedHand.push()
-      hasPushedPanCursor = true
-    }
-  }
-
-  private func updateMiddlePan(to windowLocation: NSPoint) {
-    let deltaX = windowLocation.x - middlePanStartLocation.x
-    // Window Y grows upward; canvas offset Y grows downward (SwiftUI top-left).
-    let deltaY = middlePanStartLocation.y - windowLocation.y
-    let newOffset = CGSize(
-      width: middlePanStartOffset.width + deltaX,
-      height: middlePanStartOffset.height + deltaY
-    )
-    scrollCoordinator?.setOffset(newOffset)
-  }
-
-  private func endMiddlePan() {
-    isMiddlePanning = false
-    if hasPushedPanCursor {
-      NSCursor.pop()
-      hasPushedPanCursor = false
-    }
-  }
-
-  private func tearDownMiddleButtonMonitor() {
-    if isMiddlePanning { endMiddlePan() }
-    if let monitor = middleButtonMonitor {
-      middleButtonMonitor = nil
-      DispatchQueue.main.async { MainActor.assumeIsolated { NSEvent.removeMonitor(monitor) } }
-    }
-  }
-
-  override func removeFromSuperview() {
-    tearDownMonitor()
-    tearDownMiddleButtonMonitor()
-    super.removeFromSuperview()
-  }
-}
-
-/// Screen-space transform for a card on canvas.
-private struct CardScreenGeometry {
-  var size: CGSize
-  var center: CGPoint
-  var scale: CGFloat
-}
-
-/// An `Animatable` container that interpolates a card between its in-canvas
-/// frame (`progress` 0) and the full-viewport expanded frame (`progress` 1).
-///
-/// Because `animatableData` is `progress`, SwiftUI re-evaluates `body` on every
-/// frame of the transition, so the card's size, center, and scale advance
-/// together and the terminal re-flows in lock-step with the offset/scale — a
-/// true magic-move from where the card sits. This sidesteps SwiftUI's implicit
-/// per-modifier interpolation, which only reached the Animatable terminal and
-/// left offset/scale to snap (the "grows from the center" / "no animation" bugs).
-private struct AnimatedExpandableCard<Content: View>: View, Animatable {
-  var progress: CGFloat
-  var collapsed: CardScreenGeometry
-  var expanded: CardScreenGeometry
-  let titleBarHeight: CGFloat
-  @ViewBuilder let content: (CGSize) -> Content
-
-  var animatableData: CGFloat {
-    get { progress }
-    set { progress = newValue }
-  }
-
-  var body: some View {
-    let fraction = max(0, min(1, progress))
-    let size = CGSize(
-      width: lerp(collapsed.size.width, expanded.size.width, fraction),
-      height: lerp(collapsed.size.height, expanded.size.height, fraction)
-    )
-    let center = CGPoint(
-      x: lerp(collapsed.center.x, expanded.center.x, fraction),
-      y: lerp(collapsed.center.y, expanded.center.y, fraction)
-    )
-    let scale = lerp(collapsed.scale, expanded.scale, fraction)
-    content(size)
-      .scaleEffect(scale, anchor: .center)
-      .offset(
-        x: center.x - size.width / 2,
-        y: center.y - (size.height + titleBarHeight) / 2
-      )
-  }
-
-  private func lerp(_ start: CGFloat, _ end: CGFloat, _ fraction: CGFloat) -> CGFloat {
-    start + (end - start) * fraction
   }
 }
